feat: tg bot

+++ src/modules/cron/cron.service.ts
@@ -5,6 +5,7 @@ import { PrismaService } from '@/src/core/prisma/prisma.service'
 
 import { MailService } from '../libs/mail/mail.service'
 import { S3Service } from '../libs/s3/s3.service'
+import { TelegramService } from '../libs/telegram/telegram.service'
 
 @Injectable()
 export class CronService {
@@ -12,6 +13,7 @@ export class CronService {
 		private readonly prismaService: PrismaService,
 		private readonly mailService: MailService,
 		private readonly s3Service: S3Service,
+		private readonly telegramService: TelegramService,
 	) {}
 
 	// Вызов деактивации аккаунтов каждую полночь
@@ -28,12 +30,26 @@ export class CronService {
 					lte: weekAgo,
 				},
 			},
+			include: {
+				notificationSettings: true,
+				stream: true,
+			},
 		})
 
 		for (const user of deactivatedAccounts) {
 			await this.mailService.sendAccountRemove(user.email)
 
-			this.s3Service.remove(user.avatar)
+			if (user.notificationSettings.telegramNotifications && user.telegramId) {
+				await this.telegramService.sendAccountDeletion(user.telegramId)
+			}
+
+			if (user.avatar) {
+				this.s3Service.remove(user.avatar)
+			}
+
+			if (user.stream.thumbnailUrl) {
+				this.s3Service.remove(user.stream.thumbnailUrl)
+			}
 		}
 
 		await this.prismaService.user.deleteMany({
